FA sniffer: Changes to ensure 64-bit build works
Also fix BUILD definition in makefiles for practical use.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@51164 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tools/capture.c
@@ -452,7 +452,7 @@ static void reset_progress(void)
 {
     char spaces[40];
     memset(spaces, ' ', sizeof(spaces));
-    fprintf(stderr, "%.*s\r", sizeof(spaces), spaces);
+    fprintf(stderr, "%.*s\r", (int) sizeof(spaces), spaces);
 }
